Rafactored to have on e file per logger
