+++ examples/ddpg/ddpg.py
@@ -68,7 +68,7 @@ DEFAULT_REWARD_SCALING = {
 
 @hydra.main(version_base=None, config_path=None, config_name="config")
 def main(cfg: "DictConfig"):
-    from torchrl.trainers.loggers import TensorboardLogger
+    from torchrl.trainers.loggers.tensorboard import TensorboardLogger
 
     cfg = correct_for_frame_skip(cfg)
