+++ examples/redq/redq.py
@@ -69,9 +69,7 @@ DEFAULT_REWARD_SCALING = {
 
 @hydra.main(version_base=None, config_path=None, config_name="config")
 def main(cfg: "DictConfig"):
-    from torchrl.trainers.loggers import (
-        TensorboardLogger,
-    )  # avoid loading on each process
+    from torchrl.trainers.loggers.tensorboard import TensorboardLogger
 
     cfg = correct_for_frame_skip(cfg)
