+++ setup.py
@@ -177,6 +177,7 @@ def _main():
             "tests": ["pytest", "pyyaml", "pytest-instafail"],
             "utils": [
                 "tensorboard",
+                "wandb",
                 "tqdm",
                 "hydra-core>=1.1",
                 "hydra-submitit-launcher",
